refactor(translate_apps): create a base.py file to avoid duplicating code
